Refactor files after moving fies restructure
